Add the base client and add some documentation
Creates a base client which is used to interact with the Envato API endpoints as
well as adding some documentation on how and why you should use this library.
This also sets up the CI testing using RSpec and VCR so we can :car::dash: and
have a speedy test suite to ensure all of our awesomeness is working as
expected.