Clean up error handling for non 200 HTTP requests
Improves the handling of responses that are not a HTTP 200 response. This allows
the client to be more resilent in the event that it receives a 4xx or 5xx status
code and not a hash.

Additionally, moves all the custom error classes under the `Envato` namespace to
ensure we don't hit any collisons with other gems or classes.