Use a static API version for endpoints
In it's current state, the API has some magic at the routing tier that
determines which backend to send the request to based on the version provided in
the URL which means we cannot set a global API version across the board for all
endpoints. This is being worked on and in the future will do some sane
versioning so that a global version can be provided.