Add note about not using `ENVATO_TEST_API_USERNAME` in URLs
After recreating some tests I seen that if a sanitised value (in this case
`ENVATO_TEST_API_USERNAME`) was used in the URL as a parameter, VCR would not
record the cassette and everytime you needed to pass in these details. I've
fixed this by not using the same value for `ENVATO_TEST_API_USERNAME` in my
tests where URL's are created using it.