Add `author_sales_per_month` endpoint
Create a method for authors to use in order to fetch the sales and earnings per
month. I have also made a proposal upstream in the API itself to use the correct
types for the `sales` and `earnings` fields as they are currently strings
despite them being stored and returned as the correct types pre-API mutation.